Update README.md
This repository contains a collection of Python Notebooks designed to analyze the sentiment and readability of text documents using various natural language processing techniques. The goal of this project is to provide a simple yet insightful way to gauge the sentiment distribution and readability of text content, which can be particularly useful for tasks such as content analysis, opinion mining, and understanding the complexity of written materials.